requirements-cpu.txt
tensorflow-cpu>=2.20.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
